type exchange

+++ src/kernel/interrupts/idt.c
@@ -47,7 +47,7 @@ extern void idt_flush(idt_ptr_t* ptr);
  * gn supplies numbered interrupt, while funcall is
  * function to be called when that interrupt happens.
  */
-void idt_set_gate(uint8_t gn, uint64_t funcall) {
+void idt_set_gate(uint8_t gn, uintptr_t funcall) {
     idt_gate_t* gate = &idt_entries[gn];
     gate->offset015 = funcall & 0xFFFF;
     gate->offset1631 = (funcall>>16) & 0xFFFF;
@@ -81,47 +81,47 @@ void initialize_interrupts() {
     memset(interrupt_handlers, 0, sizeof(interrupt_handlers));
 
     idt_ptr.limit = (sizeof(idt_entries)) - 1;
-    idt_ptr.base = (uint64_t) &idt_entries;
+    idt_ptr.base = (uintptr_t) &idt_entries;
 
     // GENERAL CPU INTERRUPTS
-    idt_set_gate(0, (uint64_t) isr0);
-    idt_set_gate(1, (uint64_t) isr1);
-    idt_set_gate(2, (uint64_t) isr2);
-    idt_set_gate(3, (uint64_t) isr3);
-    idt_set_gate(4, (uint64_t) isr4);
-    idt_set_gate(5, (uint64_t) isr5);
-    idt_set_gate(6, (uint64_t) isr6);
-    idt_set_gate(7, (uint64_t) isr7);
-    idt_set_gate(8, (uint64_t) isr8);
-    idt_set_gate(9, (uint64_t) isr9);
+    idt_set_gate(0, (uintptr_t) isr0);
+    idt_set_gate(1, (uintptr_t) isr1);
+    idt_set_gate(2, (uintptr_t) isr2);
+    idt_set_gate(3, (uintptr_t) isr3);
+    idt_set_gate(4, (uintptr_t) isr4);
+    idt_set_gate(5, (uintptr_t) isr5);
+    idt_set_gate(6, (uintptr_t) isr6);
+    idt_set_gate(7, (uintptr_t) isr7);
+    idt_set_gate(8, (uintptr_t) isr8);
+    idt_set_gate(9, (uintptr_t) isr9);
 
-    idt_set_gate(10, (uint64_t) isr10);
-    idt_set_gate(11, (uint64_t) isr11);
-    idt_set_gate(12, (uint64_t) isr12);
-    idt_set_gate(13, (uint64_t) isr13);
-    idt_set_gate(14, (uint64_t) isr14);
-    idt_set_gate(15, (uint64_t) isr15);
-    idt_set_gate(16, (uint64_t) isr16);
-    idt_set_gate(17, (uint64_t) isr17);
-    idt_set_gate(18, (uint64_t) isr18);
-    idt_set_gate(19, (uint64_t) isr19);
+    idt_set_gate(10, (uintptr_t) isr10);
+    idt_set_gate(11, (uintptr_t) isr11);
+    idt_set_gate(12, (uintptr_t) isr12);
+    idt_set_gate(13, (uintptr_t) isr13);
+    idt_set_gate(14, (uintptr_t) isr14);
+    idt_set_gate(15, (uintptr_t) isr15);
+    idt_set_gate(16, (uintptr_t) isr16);
+    idt_set_gate(17, (uintptr_t) isr17);
+    idt_set_gate(18, (uintptr_t) isr18);
+    idt_set_gate(19, (uintptr_t) isr19);
 
-    idt_set_gate(20, (uint64_t) isr20);
-    idt_set_gate(21, (uint64_t) isr21);
-    idt_set_gate(22, (uint64_t) isr22);
-    idt_set_gate(23, (uint64_t) isr23);
-    idt_set_gate(24, (uint64_t) isr24);
-    idt_set_gate(25, (uint64_t) isr25);
-    idt_set_gate(26, (uint64_t) isr26);
-    idt_set_gate(27, (uint64_t) isr27);
-    idt_set_gate(28, (uint64_t) isr28);
-    idt_set_gate(29, (uint64_t) isr29);
+    idt_set_gate(20, (uintptr_t) isr20);
+    idt_set_gate(21, (uintptr_t) isr21);
+    idt_set_gate(22, (uintptr_t) isr22);
+    idt_set_gate(23, (uintptr_t) isr23);
+    idt_set_gate(24, (uintptr_t) isr24);
+    idt_set_gate(25, (uintptr_t) isr25);
+    idt_set_gate(26, (uintptr_t) isr26);
+    idt_set_gate(27, (uintptr_t) isr27);
+    idt_set_gate(28, (uintptr_t) isr28);
+    idt_set_gate(29, (uintptr_t) isr29);
 
-    idt_set_gate(30, (uint64_t) isr30);
-    idt_set_gate(31, (uint64_t) isr31);
-    idt_set_gate(128, (uint64_t) isr128);
-    idt_set_gate(128, (uint64_t) isr129);
-    idt_set_gate(255, (uint64_t) isr255);
+    idt_set_gate(30, (uintptr_t) isr30);
+    idt_set_gate(31, (uintptr_t) isr31);
+    idt_set_gate(128, (uintptr_t) isr128);
+    idt_set_gate(128, (uintptr_t) isr129);
+    idt_set_gate(255, (uintptr_t) isr255);
 
     // INTERRUPTS FROM THE BOARD
     // remapping interrupts from irq
@@ -136,22 +136,22 @@ void initialize_interrupts() {
     outb(0x21, 0x0);
     outb(0xA1, 0x0);
 
-    idt_set_gate(32, (uint64_t) isr32);
-    idt_set_gate(33, (uint64_t) isr33);
-    idt_set_gate(34, (uint64_t) isr34);
-    idt_set_gate(35, (uint64_t) isr35);
-    idt_set_gate(36, (uint64_t) isr36);
-    idt_set_gate(37, (uint64_t) isr37);
-    idt_set_gate(38, (uint64_t) isr38);
-    idt_set_gate(39, (uint64_t) isr39);
-    idt_set_gate(40, (uint64_t) isr40);
-    idt_set_gate(41, (uint64_t) isr41);
-    idt_set_gate(42, (uint64_t) isr42);
-    idt_set_gate(43, (uint64_t) isr43);
-    idt_set_gate(44, (uint64_t) isr44);
-    idt_set_gate(45, (uint64_t) isr45);
-    idt_set_gate(46, (uint64_t) isr46);
-    idt_set_gate(47, (uint64_t) isr47);
+    idt_set_gate(32, (uintptr_t) isr32);
+    idt_set_gate(33, (uintptr_t) isr33);
+    idt_set_gate(34, (uintptr_t) isr34);
+    idt_set_gate(35, (uintptr_t) isr35);
+    idt_set_gate(36, (uintptr_t) isr36);
+    idt_set_gate(37, (uintptr_t) isr37);
+    idt_set_gate(38, (uintptr_t) isr38);
+    idt_set_gate(39, (uintptr_t) isr39);
+    idt_set_gate(40, (uintptr_t) isr40);
+    idt_set_gate(41, (uintptr_t) isr41);
+    idt_set_gate(42, (uintptr_t) isr42);
+    idt_set_gate(43, (uintptr_t) isr43);
+    idt_set_gate(44, (uintptr_t) isr44);
+    idt_set_gate(45, (uintptr_t) isr45);
+    idt_set_gate(46, (uintptr_t) isr46);
+    idt_set_gate(47, (uintptr_t) isr47);
 
     idt_flush(&idt_ptr);
 }
